backend_new/requirements.txt
# Sistema de Gestión Agrícola - Backend Reimplementado
# Requirements para Python 3.9+

# Framework principal
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# Base de datos
psycopg2-binary==2.9.7  # PostgreSQL adapter
SQLAlchemy==2.0.23

# Variables de entorno
python-dotenv==1.0.0

# Servidor de producción
gunicorn==21.2.0

# Utilidades
Werkzeug==2.3.7
click==8.1.7
python-dateutil==2.8.2

# Desarrollo y testing (opcional)
pytest==7.4.3
pytest-flask==1.3.0

# Seguridad
itsdangerous==2.1.2

# JSON y validación
marshmallow==3.20.1
